Revert "Migrate `token` module to Server Preset (Part 2) (#4937)"
This reverts commit bf277aea72b2c9dc9337fe96d41184183e0904f6.
