+++ .gitignore
@@ -140,4 +140,3 @@ deployment/utils/vector.types.ts
 deployment/utils/contour.types.ts
 
 schema.graphql
-resolvers.generated.ts
