+++ codegen.mts
@@ -34,6 +34,7 @@ const config: CodegenConfig = {
           'shared',
           'support',
           'target',
+          'token',
         ],
         scalarsOverrides: {
           DateTime: { type: 'string' },
@@ -55,11 +56,6 @@ const config: CodegenConfig = {
               '../modules/auth/providers/organization-access#OrganizationAccessScope',
             SupportTicketPriority: '../shared/entities#SupportTicketPriority',
             SupportTicketStatus: '../shared/entities#SupportTicketStatus',
-          },
-          resolversNonOptionalTypename: {
-            interfaceImplementingType: true,
-            unionMember: true,
-            excludeTypes: ['TokenInfoPayload'],
           },
         },
       },
