+++ packages/services/api/src/modules/token/index.ts
@@ -1,7 +1,7 @@
 import { createModule } from 'graphql-modules';
 import { TokenManager } from './providers/token-manager';
 import { TokenStorage } from './providers/token-storage';
-import { resolvers } from './resolvers.generated';
+import { resolvers } from './resolvers';
 import typeDefs from './module.graphql';
 
 export const tokenModule = createModule({
