--- packages/services/api/src/modules/token/resolvers/Target.ts
@@ -1,12 +1,0 @@
-import { TokenManager } from '../providers/token-manager';
-import type { TargetResolvers } from './../../../__generated__/types.next';
-
-export const Target: Pick<TargetResolvers, 'tokens'> = {
-  tokens(target, _, { injector }) {
-    return injector.get(TokenManager).getTokens({
-      target: target.id,
-      project: target.projectId,
-      organization: target.orgId,
-    });
-  },
-};
