--- packages/services/api/src/modules/token/resolvers/Token.ts
@@ -1,10 +1,0 @@
-import type { TokenResolvers } from './../../../__generated__/types.next';
-
-export const Token: TokenResolvers = {
-  id(token) {
-    return token.token;
-  },
-  alias(token) {
-    return token.tokenAlias;
-  },
-};
