+++ packages/services/api/src/modules/usage-estimation/resolvers.generated.ts
@@ -1,0 +1,7 @@
+/* This file was automatically generated. DO NOT UPDATE MANUALLY. */
+import type { Resolvers } from './../../__generated__/types.next';
+import { usageEstimation as Query_usageEstimation } from './resolvers/Query/usageEstimation';
+
+export const resolvers: Resolvers = {
+  Query: { usageEstimation: Query_usageEstimation },
+};
